==> 1_get_data.R <==
# 1. Parameter and  function  ----
setwd('/home/weiwewu/Uplift Model/AU/gmv_model/data')
library(readxl)
Campaign_calendar_Mega <- read_excel("Campaign_calendar_Mega.xlsx")
Megaweekend = Campaign_calendar_Mega[Campaign_calendar_Mega$campName!='SuperSunday',]


options(java.parameters="-Xmx16g")
suppressMessages(library(RJDBC))
suppressMessages(library(plyr))
suppressMessages(library(dplyr))
suppressMessages(library(stringr))

parRep = function(query) {
  query = gsub("@cohortStart", date_index[1], query, fixed=TRUE)
  query = gsub("@cohortEnd", date_index[2], query, fixed=TRUE)
  query = gsub("@measureStart", date_index[3], query, fixed=TRUE)
  query = gsub("@measureEnd", date_index[4], query, fixed=TRUE)
  return(query)
}

# k = 40
# Megaweekend[k,]

date_index = rep("2019-01-01",4)

for( k in  54: nrow(Megaweekend))
{
  cohortStart = Megaweekend[k,5]
  cohortEnd = Megaweekend[k,6]
  measureStart = Megaweekend[k,7]
  measureEnd = Megaweekend[k,8]
  
  date_index[1] = as.character(cohortStart$cohortStart)
  date_index[2] = as.character(cohortEnd$cohortEnd)
  date_index[3] = as.character(measureStart$measureStart)
  date_index[4] = as.character(measureEnd$measureEnd)
  
  dw <- config::get(value = "datawarehouse", file = "~/credentials/config.yml")
  dbCon <- dbConnect(
    drv = JDBC(
      driverClass = dw$driverClass,
      classPath = dw$classPath
    ),
    url = dw$Mozart, user = dw$user, password = dw$password
  )
  rm(dw)
  
  print(paste("################### round",k,as.Date(measureStart$measureStart) , "start #####################"))
  
  # _______________________________ ----
  # 2. Prepare data  ----
  
  # 2.1 User part  ----
  # dbSendUpdate(dbCon, "drop table AU_C2C_USR")
  
  query = "
  CREATE MULTISET VOLATILE TABLE AU_C2C_USR AS (
  SELECT
  USER_ID
  , CASE WHEN TRIM(CELLNAME) LIKE '%@_T' ESCAPE '@' THEN 'T' ELSE 'C' END AS TC_SEGMENT
  , CAST(PROCESS_DT AS DATE FORMAT'MM/DD/YYYY') AS PROMO_DT
  FROM P_WEIWEIWU_T.MEGAWEEK_BASE
  WHERE
  1 = 1
  AND PROMO_DT BETWEEN  CAST('@cohortStart' AS DATE)-5 AND CAST('@cohortEnd' AS DATE) 
  AND USER_ID < 1E15 
  AND CELLNAME NOT LIKE '%SEED%' 
  GROUP BY 1,2,3
  )
  WITH DATA
  PRIMARY INDEX (USER_ID, TC_SEGMENT, PROMO_DT)
  ON COMMIT PRESERVE ROWS
  ;
  "
  dbSendUpdate(dbCon,parRep(query))
  # dbGetQuery(dbCon, "sel tc_segment, PROMO_DT, count(distinct user_id) from AU_C2C_USR group by 1,2")
  
  query = "
  CREATE MULTISET VOLATILE TABLE AU_C2C_USR_OPTIN AS (
  SELECT
  USR.USER_ID
  ,USR.TC_SEGMENT
  ,USR.PROMO_DT
  ,CASE WHEN OPTIN.USER_ID IS NULL THEN 0 ELSE 1 END AS PROMO_OPTIN_IND
  FROM 
  AU_C2C_USR USR
  LEFT JOIN
  ACCESS_VIEWS.DW_USER_PRMTN_OPTIN OPTIN
  ON OPTIN.USER_ID = USR.USER_ID
  AND OPTIN.PRMTN_GRP_ID = 6659
  AND OPTIN.END_DT = USR.PROMO_DT + INTERVAL '2' DAY
  GROUP BY 1,2,3,4--,5,6
  )
  WITH DATA
  UNIQUE PRIMARY INDEX (USER_ID, TC_SEGMENT, PROMO_DT)
  ON COMMIT PRESERVE ROWS
  ;
  "
  dbSendUpdate(dbCon,parRep(query))
  # dbGetQuery(dbCon, "sel tc_segment, count(distinct user_id) from AU_C2C_USR_OPTIN group by 1")
  
  
  query = "
  CREATE MULTISET VOLATILE TABLE OPTIN90D_C2C AS (
  SELECT 
  U.*
  FROM AU_C2C_USR_OPTIN U 
  JOIN PRS_RESTRICTED_V.SLM_OPTIN_SEGM F12 ON U.USER_ID = F12.USER_ID 
  AND CAST('@measureStart' AS DATE)-1 BETWEEN START_DT AND END_DT
  )
  WITH DATA
  UNIQUE PRIMARY INDEX (USER_ID, TC_SEGMENT, PROMO_DT)
  ON COMMIT PRESERVE ROWS
  ;
  "
  dbSendUpdate(dbCon,parRep(query))
  # dbGetQuery(dbCon, "sel tc_segment, count(distinct user_id) from OPTIN90D_C2C group by 1")
  print(paste("round",k,as.Date(measureStart$measureStart) , "user part"))
  
  # 2.2 RESPONSE VARIABLE part  ----
  # dbSendUpdate(dbCon, "drop table AU_RESPONSE_LSTG")
  query = "
  CREATE MULTISET VOLATILE TABLE AU_RESPONSE_LSTG AS (
  SELECT
  USR.USER_ID
  ,USR.TC_SEGMENT
  ,USR.PROMO_DT
  ,LSTG_DT_AU
  ,ITEM_ID
  FROM OPTIN90D_C2C USR
  LEFT JOIN 
  (
  SEL ITEM.SLR_ID, ITEM.ITEM_ID, CAST(COLD.AUCT_START_DATE+INTERVAL'17'HOUR AS DATE) AS LSTG_DT_AU
  
  FROM DW_LSTG_ITEM ITEM 
  
  JOIN DW_LSTG_ITEM_COLD COLD ON ITEM.ITEM_ID = COLD.ITEM_ID AND ITEM.SLR_ID = COLD.SLR_ID AND ITEM.AUCT_END_DT = COLD.AUCT_END_DT
  
  JOIN DW_CATEGORY_GROUPINGS GROU ON ITEM.ITEM_SITE_ID=GROU.SITE_ID  AND ITEM.LEAF_CATEG_ID=GROU.LEAF_CATEG_ID 
  AND GROU.LEAF_CATEG_ID = GROU.MOVE_TO AND GROU.SAP_CATEGORY_ID NOT IN (5,7,23,41)
  
  WHERE 
  1 = 1
  AND ITEM.AUCT_END_DT>=DATE'2018-01-01'   
  AND CAST(COLD.AUCT_START_DATE+INTERVAL'17'HOUR AS DATE)  BETWEEN CAST('@measureStart' AS DATE) AND CAST('@measureEnd' AS DATE) + 5
  AND   ITEM.AUCT_TYPE_CODE NOT IN (12,15)           
  AND  APPLICATION_ID <>114100
  AND ITEM.SLR_ID IN  (SEL USER_ID FROM OPTIN90D_C2C GROUP BY 1)
  ) AS I ON USR.USER_ID = I.SLR_ID
  WHERE 
  1 = 1
  GROUP BY 1,2,3,4,5
  )
  WITH DATA
  UNIQUE PRIMARY INDEX (USER_ID, ITEM_ID)
  ON COMMIT PRESERVE ROWS
  ;
  "
  dbSendUpdate(dbCon,parRep(query))
  # dbGetQuery(dbCon, "sel tc_segment, count(distinct user_id), COUNT(DISTINCT ITEM_ID), COUNT(*) from AU_RESPONSE_LSTG group by 1")
  # dbGetQuery(dbCon, "sel * from AU_RESPONSE_LSTG WHERE USER_ID = 1112545210")
  # dbGetQuery(dbCon, "sel ITEM_ID, COUNT(*) AS NUM from AU_RESPONSE_LSTG GROUP BY 1 HAVING NUM>1")
  
  
  # dbSendUpdate(dbCon, "drop table AU_RESPONSE_TXN")
  query = "
  CREATE MULTISET VOLATILE TABLE AU_RESPONSE_TXN AS (
  SELECT
  USER_ID
  ,TC_SEGMENT
  ,PROMO_DT
  ,LSTG_DT_AU
  ,LIST.ITEM_ID
  ,SUM(CAST(((QUANTITY*ITEM_PRICE*LSTG_CURNCY_EXCHNG_RATE)/EXCH.EXCHNG_RATE) AS DECIMAL(18,2))) GMV
  ,ZEROIFNULL(SUM(QUANTITY)) AS SI
  FROM AU_RESPONSE_LSTG LIST
  LEFT JOIN ACCESS_VIEWS.DW_CHECKOUT_TRANS AS TRANS ON LIST.ITEM_ID = TRANS.ITEM_ID AND LIST.USER_ID = TRANS.SELLER_ID
  AND AUCT_END_DT>=DATE'2018-01-01'
  AND CAST(CREATED_TIME+INTERVAL'17'HOUR AS DATE) between CAST('@measureStart' AS DATE) AND CAST('@measureEnd' AS DATE)+30+5
  AND CK_WACKO_YN='N' AND SALE_TYPE NOT IN (12,15)        
  LEFT JOIN DW_DAILY_EXCHANGE_RATES EXCH  ON EXCH.DAY_OF_RATE_DT = TRANS.CREATED_DT AND EXCH.CURNCY_ID = 5
  WHERE 
  1 = 1
  GROUP BY 1,2,3,4,5
  )
  WITH DATA
  UNIQUE PRIMARY INDEX (USER_ID,ITEM_ID)
  ON COMMIT PRESERVE ROWS
  ;
  "
  dbSendUpdate(dbCon,parRep(query))
  # dbGetQuery(dbCon, "sel tc_segment, count(distinct user_id) from AU_RESPONSE_TXN group by 1")
  # dbGetQuery(dbCon, "sel tc_segment, count(distinct user_id), COUNT(DISTINCT ITEM_ID), COUNT(*) from AU_RESPONSE_TXN group by 1")
  
  # dbSendUpdate(dbCon, "drop table AU_RESPONSE")
  query = "
  CREATE MULTISET VOLATILE TABLE AU_RESPONSE AS (
  SELECT
  A.USER_ID
  ,A.TC_SEGMENT
  ,A.PROMO_DT
  ,COUNT(DISTINCT CASE WHEN LSTG_DT_AU between CAST('@measureStart' AS DATE) AND CAST('@measureEnd' AS DATE) THEN ITEM_ID END) AS promo_lstg_cnt_2d
  ,COUNT(DISTINCT CASE WHEN LSTG_DT_AU between CAST('@measureStart' AS DATE) AND CAST('@measureEnd' AS DATE)+5 THEN ITEM_ID END) AS promo_lstg_cnt_7d
  ,SUM(CASE WHEN LSTG_DT_AU between CAST('@measureStart' AS DATE) AND CAST('@measureEnd' AS DATE) THEN GMV END) AS promo_gmv_2d
  ,SUM(CASE WHEN LSTG_DT_AU between CAST('@measureStart' AS DATE) AND CAST('@measureEnd' AS DATE)+5 THEN GMV END) AS promo_gmv_7d
  ,SUM(CASE WHEN LSTG_DT_AU between CAST('@measureStart' AS DATE) AND CAST('@measureEnd' AS DATE) THEN SI END) AS promo_si_2d
  ,SUM(CASE WHEN LSTG_DT_AU between CAST('@measureStart' AS DATE) AND CAST('@measureEnd' AS DATE)+5 THEN SI END) AS promo_si_7d
  FROM AU_RESPONSE_TXN  A
  group by 1,2,3
  )
  WITH DATA
  UNIQUE PRIMARY INDEX (USER_ID)
  ON COMMIT PRESERVE ROWS
  ;
  "
  dbSendUpdate(dbCon,parRep(query))
  # dbGetQuery(dbCon, "sel tc_segment, count(distinct user_id), SUM(promo_lstg_cnt_7d) from AU_RESPONSE group by 1")
  # dbGetQuery(dbCon, "sel * from AU_RESPONSE where promo_lstg_cnt>0")
  
  print(paste("round",k,as.Date(measureStart$measureStart) , "response part"))
  
  
  # 2.3 FEATURES part  ----
  
  # 2.31 OPT-IN INFO  ----
  
  query = "
  CREATE MULTISET VOLATILE TABLE AU_PREV_OPTINS AS (
  SELECT
  USR.USER_ID
  ,USR.PROMO_DT
  ,SUM(CASE WHEN OPTIN.USER_ID IS NULL THEN 0 ELSE 1 END) PREV_OPTIN_CNT
  FROM 
  OPTIN90D_C2C USR
  LEFT JOIN
  ACCESS_VIEWS.DW_USER_PRMTN_OPTIN OPTIN
  ON OPTIN.USER_ID = USR.USER_ID
  -- PROMOS PREVIOUS TO USR.PROMO_DT AND AFTER '2017-01-01'
  AND OPTIN.END_DT BETWEEN PROMO_DT - INTERVAL '730' DAY AND PROMO_DT - INTERVAL '1' DAY
  GROUP BY 1,2
  ) 
  WITH DATA
  UNIQUE PRIMARY INDEX (USER_ID, PROMO_DT)
  ON COMMIT PRESERVE ROWS
  ;
  "
  dbSendUpdate(dbCon,parRep(query))
  
  # 2.32 LISTING  ----
  
  query = "
  CREATE MULTISET VOLATILE TABLE AU_LISTINGS AS (
  SELECT
  USR.USER_ID
  ,USR.PROMO_DT
  ,SUM(CASE WHEN SD.CAL_DT >= USR.PROMO_DT - INTERVAL  '30' DAY THEN NEW_LSTG_CNT ELSE 0 END) AS T30_LSTG_CNT
  ,SUM(CASE WHEN SD.CAL_DT >= USR.PROMO_DT - INTERVAL  '90' DAY THEN NEW_LSTG_CNT ELSE 0 END) AS T90_LSTG_CNT	
  ,SUM(CASE WHEN SD.CAL_DT >= USR.PROMO_DT - INTERVAL '180' DAY THEN NEW_LSTG_CNT ELSE 0 END) AS T180_LSTG_CNT
  ,SUM(CASE WHEN SD.CAL_DT >= USR.PROMO_DT - INTERVAL '365' DAY THEN NEW_LSTG_CNT ELSE 0 END) AS T365_LSTG_CNT
  -- SAME TIME WINDOW BUT SHIFTED TO ONE PERIOD BEFORE
  ,SUM(CASE WHEN SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL  '60' DAY AND USR.PROMO_DT - INTERVAL  '31' DAY THEN NEW_LSTG_CNT ELSE 0 END) AS TPREV30_LSTG_CNT
  ,SUM(CASE WHEN SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL '180' DAY AND USR.PROMO_DT - INTERVAL  '91' DAY THEN NEW_LSTG_CNT ELSE 0 END) AS TPREV90_LSTG_CNT	
  ,SUM(CASE WHEN SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL '360' DAY AND USR.PROMO_DT - INTERVAL '181' DAY THEN NEW_LSTG_CNT ELSE 0 END) AS TPREV180_LSTG_CNT
  ,SUM(CASE WHEN SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL '730' DAY AND USR.PROMO_DT - INTERVAL '366' DAY THEN NEW_LSTG_CNT ELSE 0 END) AS TPREV365_LSTG_CNT
  -- CHANGE RATES OVER THE SAME TIME WINDOW
  ,CASE WHEN TPREV30_LSTG_CNT <> 0 THEN CAST(T30_LSTG_CNT AS DECIMAL(18,2)) / TPREV30_LSTG_CNT - 1 
        WHEN T30_LSTG_CNT > 0 THEN 1 ELSE 0  END AS T30_LSTG_RATE
  ,CASE WHEN TPREV90_LSTG_CNT <> 0 THEN CAST(T90_LSTG_CNT AS DECIMAL(18,2)) / TPREV90_LSTG_CNT - 1 
        WHEN T90_LSTG_CNT > 0 THEN 1 ELSE 0  END AS T90_LSTG_RATE
  ,CASE WHEN TPREV180_LSTG_CNT <> 0 THEN CAST(T180_LSTG_CNT AS DECIMAL(18,2)) / TPREV180_LSTG_CNT - 1 
        WHEN T180_LSTG_CNT > 0 THEN 1 ELSE 0  END AS T180_LSTG_RATE
  ,CASE WHEN TPREV365_LSTG_CNT <> 0 THEN CAST(T365_LSTG_CNT AS DECIMAL(18,2)) / TPREV365_LSTG_CNT - 1 
        WHEN T365_LSTG_CNT > 0 THEN 1 ELSE 0  END AS T365_LSTG_RATE
  -- LISTING TOOLS USED IN THE YEAR BEFORE THE PROMO
  ,MAX(CASE WHEN SD.CAL_DT >= USR.PROMO_DT - INTERVAL '365' DAY AND SD.SLNG_TOOL_CD IN (13,2,9,14,6,3,10) THEN 1 ELSE 0 END) AS LSTG_TOOL_PRO_IND   --EasyLister / File Exchange / TurboLister / Inkfrog / Auctiva / Selling manager pr / BEAR
  ,MAX(CASE WHEN SD.CAL_DT >= USR.PROMO_DT - INTERVAL '365' DAY AND SD.SLNG_TOOL_CD IN (15,19) THEN 1 ELSE 0 END) AS LSTG_TOOL_IOS_IND
  ,MAX(CASE WHEN SD.CAL_DT >= USR.PROMO_DT - INTERVAL '365' DAY AND SD.SLNG_TOOL_CD IN (17) THEN 1 ELSE 0 END) AS LSTG_TOOL_ANDROID_IND
  ,MAX(CASE WHEN SD.CAL_DT >= USR.PROMO_DT - INTERVAL '365' DAY AND SD.SLNG_TOOL_CD IN (15,17,19,23) THEN 1 ELSE 0 END) AS LSTG_TOOL_MOB_IND
  FROM
  OPTIN90D_C2C USR
  LEFT JOIN
  PRS_RESTRICTED_V.SLR_INVNTRY_SD SD
  ON USR.USER_ID = SD.SLR_ID
  -- LISTINGS INSERTED BEFORE THE PROMO WINDOW, GOING BACK AT MOST TWO YEARS
  AND SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL '730' DAY AND  USR.PROMO_DT - INTERVAL '1' DAY
  -- EARLIEST PROMO IS IN JAN 2019, SO TWO YEARS BEFORE THAT
  AND SD.CAL_DT >= DATE '2017-01-01'
  -- RESTRICT TO DE LISTINGS FOR EASIER JOINING TO THE CATEGORY TABLE
  AND SD.LSTG_SITE_ID = 15
  GROUP BY 1,2
  )
  WITH DATA
  UNIQUE PRIMARY INDEX (USER_ID, PROMO_DT)
  ON COMMIT PRESERVE ROWS
  ;
  "
  dbSendUpdate(dbCon,parRep(query))
  
  print(paste("round",k,as.Date(measureStart$measureStart) , "listing part"))
  
  # 2.33 VERTICAL  ----
  
  query = "
  CREATE MULTISET VOLATILE TABLE CAT AS
  (
  SELECT
  LEAF_CATEG_ID
  ,BSNS_VRTCL_NAME
  FROM
  ACCESS_VIEWS.DW_CATEGORY_GROUPINGS
  WHERE
  SITE_ID IN (15)
  AND MOVE_TO = LEAF_CATEG_ID
  )
  WITH DATA
  PRIMARY INDEX(LEAF_CATEG_ID)
  ON COMMIT PRESERVE ROWS;
  "
  dbSendUpdate(dbCon,parRep(query))
  
  query = "
  COLLECT STATISTICS INDEX (LEAF_CATEG_ID) ON CAT;
  "
  dbSendUpdate(dbCon,parRep(query))
  
  query = "
  CREATE MULTISET VOLATILE TABLE AU_VERTICALS AS (
  SELECT
  USR.USER_ID
  ,USR.PROMO_DT
  ,COUNT(DISTINCT BSNS_VRTCL_NAME) AS T365_VRTCL_CNT
  FROM
  OPTIN90D_C2C USR
  LEFT JOIN
  PRS_RESTRICTED_V.SLR_INVNTRY_SD SD
  ON USR.USER_ID = SD.SLR_ID
  -- LISTINGS INSERTED BEFORE THE PROMO WINDOW, GOING BACK AT MOST TWO YEARS
  AND SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL '365' DAY AND  USR.PROMO_DT - INTERVAL '1' DAY
  -- EARLIEST PROMO IS IN JAN 2019, SO ONE YEARS BEFORE THAT
  AND SD.CAL_DT >= DATE '2017-01-01'
  -- RESTRICT TO DE LISTINGS FOR EASIER JOINING TO THE CATEGORY TABLE
  AND SD.LSTG_SITE_ID = 15	
  LEFT JOIN
  CAT
  ON SD.LEAF_CATEG_ID = CAT.LEAF_CATEG_ID
  GROUP BY 1,2
  )
  WITH DATA
  UNIQUE PRIMARY INDEX (USER_ID, PROMO_DT)
  ON COMMIT PRESERVE ROWS
  ;
  
  "
  dbSendUpdate(dbCon,parRep(query))
  
  print(paste("round",k,as.Date(measureStart$measureStart) , "vertical part"))
  
  # 2.34 SELL  ----
  
  query = "
  CREATE MULTISET VOLATILE TABLE AU_SELLS AS (
  SELECT
  USR.USER_ID
  ,USR.PROMO_DT
  ,SUM(CASE WHEN SD.CAL_DT >= USR.PROMO_DT - INTERVAL  '30' DAY THEN SOLD_ITEM_CNT ELSE 0 END) AS T30_SI_CNT
  ,SUM(CASE WHEN SD.CAL_DT >= USR.PROMO_DT - INTERVAL  '90' DAY THEN SOLD_ITEM_CNT ELSE 0 END) AS T90_SI_CNT
  ,SUM(CASE WHEN SD.CAL_DT >= USR.PROMO_DT - INTERVAL '180' DAY THEN SOLD_ITEM_CNT ELSE 0 END) AS T180_SI_CNT
  ,SUM(CASE WHEN SD.CAL_DT >= USR.PROMO_DT - INTERVAL '365' DAY THEN SOLD_ITEM_CNT ELSE 0 END) AS T365_SI_CNT
  -- SAME TIME WINDOW BUT SHIFTED TO ONE PERIOD BEFORE
  ,SUM(CASE WHEN SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL  '60' DAY AND USR.PROMO_DT - INTERVAL  '31' DAY THEN SOLD_ITEM_CNT ELSE 0 END) AS TPREV30_SI_CNT
  ,SUM(CASE WHEN SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL '180' DAY AND USR.PROMO_DT - INTERVAL  '91' DAY THEN SOLD_ITEM_CNT ELSE 0 END) AS TPREV90_SI_CNT
  ,SUM(CASE WHEN SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL '360' DAY AND USR.PROMO_DT - INTERVAL '181' DAY THEN SOLD_ITEM_CNT ELSE 0 END) AS TPREV180_SI_CNT
  ,SUM(CASE WHEN SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL '730' DAY AND USR.PROMO_DT - INTERVAL '366' DAY THEN SOLD_ITEM_CNT ELSE 0 END) AS TPREV365_SI_CNT
  -- CHANGE RATES OVER THE SAME TIME WINDOW
  ,CASE WHEN TPREV30_SI_CNT <> 0 THEN CAST(T30_SI_CNT AS DECIMAL(18,2)) / TPREV30_SI_CNT - 1 
        WHEN T30_SI_CNT > 0 THEN 1 ELSE 0  END AS T30_SI_RATE
  ,CASE WHEN TPREV90_SI_CNT <> 0 THEN CAST(T90_SI_CNT AS DECIMAL(18,2)) / TPREV90_SI_CNT - 1 
        WHEN T90_SI_CNT > 0 THEN 1 ELSE 0  END AS T90_SI_RATE
  ,CASE WHEN TPREV180_SI_CNT <> 0 THEN CAST(T180_SI_CNT AS DECIMAL(18,2)) / TPREV180_SI_CNT - 1 
        WHEN T180_SI_CNT > 0 THEN 1 ELSE 0  END AS T180_SI_RATE
  ,CASE WHEN TPREV365_SI_CNT <> 0 THEN CAST(T365_SI_CNT AS DECIMAL(18,2)) / TPREV365_SI_CNT - 1 
        WHEN T365_SI_CNT > 0 THEN 1 ELSE 0  END AS T365_SI_RATE
  ,SUM(CASE WHEN SD.CAL_DT >= USR.PROMO_DT - INTERVAL  '30' DAY THEN GMV_USD_AMT ELSE 0 END) AS T30_GMV_AMT
  ,SUM(CASE WHEN SD.CAL_DT >= USR.PROMO_DT - INTERVAL  '90' DAY THEN GMV_USD_AMT ELSE 0 END) AS T90_GMV_AMT
  ,SUM(CASE WHEN SD.CAL_DT >= USR.PROMO_DT - INTERVAL '180' DAY THEN GMV_USD_AMT ELSE 0 END) AS T180_GMV_AMT
  ,SUM(CASE WHEN SD.CAL_DT >= USR.PROMO_DT - INTERVAL '365' DAY THEN GMV_USD_AMT ELSE 0 END) AS T365_GMV_AMT
  -- SAME TIME WINDOW BUT SHIFTED TO ONE PERIOD BEFORE
  ,SUM(CASE WHEN SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL  '60' DAY AND USR.PROMO_DT - INTERVAL  '31' DAY THEN GMV_USD_AMT ELSE 0 END) AS TPREV30_GMV_AMT
  ,SUM(CASE WHEN SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL '180' DAY AND USR.PROMO_DT - INTERVAL  '91' DAY THEN GMV_USD_AMT ELSE 0 END) AS TPREV90_GMV_AMT
  ,SUM(CASE WHEN SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL '360' DAY AND USR.PROMO_DT - INTERVAL '181' DAY THEN GMV_USD_AMT ELSE 0 END) AS TPREV180_GMV_AMT
  ,SUM(CASE WHEN SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL '730' DAY AND USR.PROMO_DT - INTERVAL '366' DAY THEN GMV_USD_AMT ELSE 0 END) AS TPREV365_GMV_AMT
  -- CHANGE RATES OVER THE SAME TIME WINDOW
  ,CASE WHEN TPREV30_GMV_AMT <> 0 THEN CAST(T30_GMV_AMT AS DECIMAL(18,2)) / TPREV30_GMV_AMT - 1 WHEN T30_GMV_AMT > 0 THEN 1 ELSE 0  END AS T30_GMV_RATE
  ,CASE WHEN TPREV90_GMV_AMT <> 0 THEN CAST(T90_GMV_AMT AS DECIMAL(18,2)) / TPREV90_GMV_AMT - 1 WHEN T90_GMV_AMT > 0 THEN 1 ELSE 0  END AS T90_GMV_RATE
  ,CASE WHEN TPREV180_GMV_AMT <> 0 THEN CAST(T180_GMV_AMT AS DECIMAL(18,2)) / TPREV180_GMV_AMT - 1 WHEN T180_GMV_AMT > 0 THEN 1 ELSE 0  END AS T180_GMV_RATE
  ,CASE WHEN TPREV365_GMV_AMT <> 0 THEN CAST(T365_GMV_AMT AS DECIMAL(18,2)) / TPREV365_GMV_AMT - 1 WHEN T365_GMV_AMT > 0 THEN 1 ELSE 0  END AS T365_GMV_RATE
  -- ASP
  ,CASE WHEN T30_SI_CNT > 0 THEN T30_GMV_AMT / T30_SI_CNT ELSE 0 END AS T30_ASP
  ,CASE WHEN T90_SI_CNT > 0 THEN T90_GMV_AMT / T90_SI_CNT ELSE 0 END AS T90_ASP
  ,CASE WHEN T180_SI_CNT > 0 THEN T180_GMV_AMT / T180_SI_CNT ELSE 0 END AS T180_ASP
  ,CASE WHEN T365_SI_CNT > 0 THEN T365_GMV_AMT / T365_SI_CNT ELSE 0 END AS T365_ASP
  FROM
  OPTIN90D_C2C USR
  LEFT JOIN
  PRS_RESTRICTED_V.SLR_TXN_SD SD
  ON USR.USER_ID = SD.SLR_ID
  -- LISTINGS INSERTED BEFORE THE PROMO WINDOW, GOING BACK AT MOST TWO YEARS
  AND SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL '730' DAY AND  USR.PROMO_DT - INTERVAL '1' DAY
  -- EARLIEST PROMO IS IN JAN 2019, SO TWO YEARS BEFORE THAT
  AND SD.CAL_DT >= DATE '2017-01-01'
  GROUP BY 1,2
  )
  WITH DATA
  UNIQUE PRIMARY INDEX (USER_ID, PROMO_DT)
  ON COMMIT PRESERVE ROWS
  ;
  "
  dbSendUpdate(dbCon,parRep(query))
  
  print(paste("round",k,as.Date(measureStart$measureStart) , "selling part"))
  
  # 2.35 EMAIL  ----
  
  query = "
  CREATE MULTISET VOLATILE TABLE EMAIL_BASE AS (
  SELECT
  CMC.USER_ID 
  ,CMC.CMPGN_RUN_DT
  ,COUNT(*) AS EMAIL_SENT_CNT
  ,SUM(CASE WHEN CMC.OPEN_CNT >= 1 THEN 1 ELSE 0 END) AS EMAIL_OPEN_CNT
  ,SUM(CASE WHEN CMC.CLICKED_COUNT >= 1 THEN 1 ELSE 0 END) AS EMAIL_CLICK_CNT
  FROM
  ACCESS_VIEWS.DW_CMC_CNTCT CMC
  WHERE
  CMC.USER_ID IN (SELECT DISTINCT USER_ID FROM OPTIN90D_C2C)
  AND CMC.CMPGN_RUN_DT >= DATE '2018-01-01'
  GROUP BY 1,2
  )
  WITH DATA AND STATISTICS
  PRIMARY INDEX (USER_ID)
  ON COMMIT PRESERVE ROWS
  ;
  "
  dbSendUpdate(dbCon,parRep(query))
  
  query = "
  CREATE MULTISET VOLATILE TABLE AU_EMAIL AS (
  SELECT
  USR.USER_ID
  ,USR.PROMO_DT
  -- EMAIL ENGAGAMENT JUST BEFORE PROMO
  -- SENT
  ,SUM(CASE WHEN CMC.CMPGN_RUN_DT BETWEEN USR.PROMO_DT - INTERVAL '90' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN CMC.EMAIL_SENT_CNT ELSE 0 END) AS T90_EMAIL_SENT_CNT
  ,SUM(CASE WHEN CMC.CMPGN_RUN_DT BETWEEN USR.PROMO_DT - INTERVAL '180' DAY AND USR.PROMO_DT - INTERVAL '91' DAY THEN CMC.EMAIL_SENT_CNT ELSE 0 END) AS TPREV90_EMAIL_SENT_CNT
  ,SUM(CASE WHEN CMC.CMPGN_RUN_DT BETWEEN USR.PROMO_DT - INTERVAL '30' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN CMC.EMAIL_SENT_CNT ELSE 0 END) AS T30_EMAIL_SENT_CNT
  ,SUM(CASE WHEN CMC.CMPGN_RUN_DT BETWEEN USR.PROMO_DT - INTERVAL '60' DAY AND USR.PROMO_DT - INTERVAL '31' DAY THEN CMC.EMAIL_SENT_CNT ELSE 0 END) AS TPREV30_EMAIL_SENT_CNT
  -- OPENED
  ,SUM(CASE WHEN CMC.CMPGN_RUN_DT BETWEEN USR.PROMO_DT - INTERVAL '90' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN CMC.EMAIL_OPEN_CNT ELSE 0 END) AS T90_EMAIL_OPEN_CNT
  ,SUM(CASE WHEN CMC.CMPGN_RUN_DT BETWEEN USR.PROMO_DT - INTERVAL '180' DAY AND USR.PROMO_DT - INTERVAL '91' DAY THEN CMC.EMAIL_OPEN_CNT ELSE 0 END) AS TPREV90_EMAIL_OPEN_CNT
  ,SUM(CASE WHEN CMC.CMPGN_RUN_DT BETWEEN USR.PROMO_DT - INTERVAL '30' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN CMC.EMAIL_OPEN_CNT ELSE 0 END) AS T30_EMAIL_OPEN_CNT
  ,SUM(CASE WHEN CMC.CMPGN_RUN_DT BETWEEN USR.PROMO_DT - INTERVAL '60' DAY AND USR.PROMO_DT - INTERVAL '31' DAY THEN CMC.EMAIL_OPEN_CNT ELSE 0 END) AS TPREV30_EMAIL_OPEN_CNT
  -- CLICKED
  ,SUM(CASE WHEN CMC.CMPGN_RUN_DT BETWEEN USR.PROMO_DT - INTERVAL '90' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN CMC.EMAIL_CLICK_CNT ELSE 0 END) AS T90_EMAIL_CLICK_CNT
  ,SUM(CASE WHEN CMC.CMPGN_RUN_DT BETWEEN USR.PROMO_DT - INTERVAL '180' DAY AND USR.PROMO_DT - INTERVAL '91' DAY THEN CMC.EMAIL_CLICK_CNT ELSE 0 END) AS TPREV90_EMAIL_CLICK_CNT
  ,SUM(CASE WHEN CMC.CMPGN_RUN_DT BETWEEN USR.PROMO_DT - INTERVAL '30' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN CMC.EMAIL_CLICK_CNT ELSE 0 END) AS T30_EMAIL_CLICK_CNT
  ,SUM(CASE WHEN CMC.CMPGN_RUN_DT BETWEEN USR.PROMO_DT - INTERVAL '60' DAY AND USR.PROMO_DT - INTERVAL '31' DAY THEN CMC.EMAIL_CLICK_CNT ELSE 0 END) AS TPREV30_EMAIL_CLICK_CNT
  -- EMAIL OPEN AND CLICK-TO-OPEN RATES 
  ,CASE WHEN T90_EMAIL_SENT_CNT = 0 THEN 0 ELSE CAST(T90_EMAIL_OPEN_CNT AS DECIMAL(18,2)) / T90_EMAIL_SENT_CNT END AS T90_EMAIL_OR
  ,CASE WHEN T30_EMAIL_SENT_CNT = 0 THEN 0 ELSE CAST(T30_EMAIL_OPEN_CNT AS DECIMAL(18,2)) / T30_EMAIL_SENT_CNT END AS T30_EMAIL_OR
  ,CASE WHEN T90_EMAIL_OPEN_CNT = 0 THEN 0 ELSE CAST(T90_EMAIL_CLICK_CNT AS DECIMAL(18,2)) / T90_EMAIL_OPEN_CNT END AS T90_EMAIL_CTOR
  ,CASE WHEN T30_EMAIL_OPEN_CNT = 0 THEN 0 ELSE CAST(T30_EMAIL_CLICK_CNT AS DECIMAL(18,2)) / T30_EMAIL_OPEN_CNT END AS T30_EMAIL_CTOR
  -- CHANGE RATES OF EMAIL OPENS AND CLICKS
  ,CASE WHEN TPREV90_EMAIL_OPEN_CNT <> 0 THEN CAST(T90_EMAIL_OPEN_CNT AS DECIMAL(18,2)) / TPREV90_EMAIL_OPEN_CNT - 1 
        WHEN T90_EMAIL_OPEN_CNT > 0 THEN 1 ELSE 0 END AS T90_EMAIL_OPENS_RATE
  ,CASE WHEN TPREV30_EMAIL_OPEN_CNT <> 0 THEN CAST(T30_EMAIL_OPEN_CNT AS DECIMAL(18,2)) / TPREV30_EMAIL_OPEN_CNT - 1 
        WHEN T30_EMAIL_OPEN_CNT > 0 THEN 1 ELSE 0 END AS T30_EMAIL_OPENS_RATE
  ,CASE WHEN TPREV90_EMAIL_CLICK_CNT <> 0 THEN CAST(T90_EMAIL_CLICK_CNT AS DECIMAL(18,2)) / TPREV90_EMAIL_CLICK_CNT - 1 
        WHEN T90_EMAIL_CLICK_CNT > 0 THEN 1 ELSE 0 END AS T90_EMAIL_CLICKS_RATE
  ,CASE WHEN TPREV30_EMAIL_CLICK_CNT <> 0 THEN CAST(T30_EMAIL_CLICK_CNT AS DECIMAL(18,2)) / TPREV30_EMAIL_CLICK_CNT - 1 
        WHEN T30_EMAIL_CLICK_CNT > 0 THEN 1 ELSE 0 END AS T30_EMAIL_CLICKS_RATE
  FROM
  OPTIN90D_C2C USR
  LEFT JOIN
  EMAIL_BASE CMC
  ON USR.USER_ID = CMC.USER_ID
  GROUP BY 1,2
  )
  WITH DATA
  UNIQUE PRIMARY INDEX (USER_ID, PROMO_DT)
  ON COMMIT PRESERVE ROWS
  ;
  "
  dbSendUpdate(dbCon,parRep(query))
  
  print(paste("round",k,as.Date(measureStart$measureStart) , "email part"))
  
  # 2.36 SEARCH  ----
  
  query = "
  CREATE MULTISET VOLATILE TABLE AU_SEARCH_BROWSE AS (
  SELECT
  USR.USER_ID
  ,USR.PROMO_DT
  -- BROWSES
  ,SUM(CASE WHEN SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL '90' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN SD.VI_PAGE_CNT ELSE 0 END) AS T90_VI_CNT
  ,SUM(CASE WHEN SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL '180' DAY AND USR.PROMO_DT - INTERVAL '91' DAY THEN SD.VI_PAGE_CNT ELSE 0 END) AS TPREV90_VI_CNT
  ,SUM(CASE WHEN SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL '30' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN SD.VI_PAGE_CNT ELSE 0 END) AS T30_VI_CNT
  ,SUM(CASE WHEN SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL '60' DAY AND USR.PROMO_DT - INTERVAL '31' DAY THEN SD.VI_PAGE_CNT ELSE 0 END) AS TPREV30_VI_CNT
  -- SEARCHES
  ,SUM(CASE WHEN SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL '90' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN SD.SRCH_PAGE_CNT ELSE 0 END) AS T90_SRCH_CNT
  ,SUM(CASE WHEN SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL '180' DAY AND USR.PROMO_DT - INTERVAL '91' DAY THEN SD.SRCH_PAGE_CNT ELSE 0 END) AS TPREV90_SRCH_CNT
  ,SUM(CASE WHEN SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL '30' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN SD.SRCH_PAGE_CNT ELSE 0 END) AS T30_SRCH_CNT
  ,SUM(CASE WHEN SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL '60' DAY AND USR.PROMO_DT - INTERVAL '31' DAY THEN SD.SRCH_PAGE_CNT ELSE 0 END) AS TPREV30_SRCH_CNT
  -- CHANGE RATES BROWSES & SEARCHES
  ,CASE WHEN TPREV90_VI_CNT <> 0 THEN CAST(T90_VI_CNT AS DECIMAL(18,2)) / TPREV90_VI_CNT - 1 
        WHEN T90_VI_CNT > 0 THEN 1 ELSE 0 END AS T90_VI_RATE
  ,CASE WHEN TPREV30_VI_CNT <> 0 THEN CAST(T30_VI_CNT AS DECIMAL(18,2)) / TPREV30_VI_CNT - 1 
        WHEN T30_VI_CNT > 0 THEN 1 ELSE 0 END AS T30_VI_RATE
  ,CASE WHEN TPREV90_SRCH_CNT <> 0 THEN CAST(T90_SRCH_CNT AS DECIMAL(18,2)) / TPREV90_SRCH_CNT - 1 
        WHEN T90_SRCH_CNT > 0 THEN 1 ELSE 0 END AS T90_SRCH_RATE
  ,CASE WHEN TPREV30_SRCH_CNT <> 0 THEN CAST(T30_SRCH_CNT AS DECIMAL(18,2)) / TPREV30_SRCH_CNT - 1 
        WHEN T30_SRCH_CNT > 0 THEN 1 ELSE 0 END AS T30_SRCH_RATE
  FROM
  OPTIN90D_C2C USR
  LEFT JOIN
  PRS_RESTRICTED_V.USER_DNA_BHVR_SD SD
  ON USR.USER_ID = SD.USER_ID
  AND SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL '180' DAY AND  USR.PROMO_DT - INTERVAL '1' DAY
  -- EARLIEST PROMO IS IN JAN 2019, SO A YEARS BEFORE THAT
  AND SD.CAL_DT >= DATE '2018-01-01'
  GROUP BY 1,2
  )
  WITH DATA
  UNIQUE PRIMARY INDEX (USER_ID, PROMO_DT)
  ON COMMIT PRESERVE ROWS
  ;
  "
  dbSendUpdate(dbCon,parRep(query))
  
  print(paste("round",k,as.Date(measureStart$measureStart) , "search part"))
  
  # 2.37 BUY  ----
  
  query = "
  CREATE MULTISET VOLATILE TABLE AU_BUYS AS (
  SELECT
  USR.USER_ID
  ,USR.PROMO_DT
  -- PURCHASES
  ,SUM(CASE WHEN SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL '365' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN SD.ITEM_PRCHSD_CNT ELSE 0 END) AS T365_PRCHS_CNT
  ,SUM(CASE WHEN SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL '730' DAY AND USR.PROMO_DT - INTERVAL '366' DAY THEN SD.ITEM_PRCHSD_CNT ELSE 0 END) AS TPREV365_PRCHS_CNT
  ,SUM(CASE WHEN SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL '90' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN SD.ITEM_PRCHSD_CNT ELSE 0 END) AS T90_PRCHS_CNT
  ,SUM(CASE WHEN SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL '180' DAY AND USR.PROMO_DT - INTERVAL '91' DAY THEN SD.ITEM_PRCHSD_CNT ELSE 0 END) AS TPREV90_PRCHS_CNT
  ,SUM(CASE WHEN SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL '30' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN SD.ITEM_PRCHSD_CNT ELSE 0 END) AS T30_PRCHS_CNT
  ,SUM(CASE WHEN SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL '60' DAY AND USR.PROMO_DT - INTERVAL '31' DAY THEN SD.ITEM_PRCHSD_CNT ELSE 0 END) AS TPREV30_PRCHS_CNT
  -- GMB
  ,SUM(CASE WHEN SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL '365' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN SD.GMB_USD_AMT ELSE 0 END) AS T365_GMB_AMT
  ,SUM(CASE WHEN SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL '730' DAY AND USR.PROMO_DT - INTERVAL '366' DAY THEN SD.GMB_USD_AMT ELSE 0 END) AS TPREV365_GMB_AMT
  ,SUM(CASE WHEN SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL '90' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN SD.GMB_USD_AMT ELSE 0 END) AS T90_GMB_AMT
  ,SUM(CASE WHEN SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL '180' DAY AND USR.PROMO_DT - INTERVAL '91' DAY THEN SD.GMB_USD_AMT ELSE 0 END) AS TPREV90_GMB_AMT
  ,SUM(CASE WHEN SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL '30' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN SD.GMB_USD_AMT ELSE 0 END) AS T30_GMB_AMT
  ,SUM(CASE WHEN SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL '60' DAY AND USR.PROMO_DT - INTERVAL '31' DAY THEN SD.GMB_USD_AMT ELSE 0 END) AS TPREV30_GMB_AMT
  -- CHANGE RATES PURCHASES & GMB
  ,CASE WHEN TPREV365_PRCHS_CNT <> 0 THEN CAST(T365_PRCHS_CNT AS DECIMAL(18,2)) / TPREV365_PRCHS_CNT - 1 
        WHEN T365_PRCHS_CNT > 0 THEN 1 ELSE 0 END AS T365_PRCHS_RATE
  ,CASE WHEN TPREV90_PRCHS_CNT <> 0 THEN CAST(T90_PRCHS_CNT AS DECIMAL(18,2)) / TPREV90_PRCHS_CNT - 1 
        WHEN T90_PRCHS_CNT > 0 THEN 1 ELSE 0 END AS T90_PRCHS_RATE
  ,CASE WHEN TPREV30_PRCHS_CNT <> 0 THEN CAST(T30_PRCHS_CNT AS DECIMAL(18,2)) / TPREV30_PRCHS_CNT - 1 
        WHEN T30_PRCHS_CNT > 0 THEN 1 ELSE 0 END AS T30_PRCHS_RATE
  ,CASE WHEN TPREV365_GMB_AMT <> 0 THEN CAST(T365_GMB_AMT AS DECIMAL(18,2)) / TPREV365_GMB_AMT - 1 
        WHEN T365_GMB_AMT > 0 THEN 1 ELSE 0 END AS T365_GMB_RATE
  ,CASE WHEN TPREV90_GMB_AMT <> 0 THEN CAST(T90_GMB_AMT AS DECIMAL(18,2)) / TPREV90_GMB_AMT - 1 
        WHEN T90_GMB_AMT > 0 THEN 1 ELSE 0 END AS T90_GMB_RATE
  ,CASE WHEN TPREV30_GMB_AMT <> 0 THEN CAST(T30_GMB_AMT AS DECIMAL(18,2)) / TPREV30_GMB_AMT - 1 
        WHEN T30_GMB_AMT > 0 THEN 1 ELSE 0 END AS T30_GMB_RATE
  FROM
  OPTIN90D_C2C USR
  LEFT JOIN
  PRS_RESTRICTED_V.USER_DNA_BBOWACL_SD SD
  ON USR.USER_ID = SD.USER_ID
  -- ONLY PURCHASE ENTRIES FROM USER_DNA_SD 
  AND SD.ITEM_PRCHSD_CNT > 0
  -- MARKETPLACES ONLY (AUCTION + FP)
  AND SD.LSTG_TYPE_CD IN (1,7,9)
  AND SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL '730' DAY AND  USR.PROMO_DT - INTERVAL '1' DAY
  -- EARLIEST PROMO IS IN JAN 2019, SO TWO YEARS BEFORE THAT
  AND SD.CAL_DT >= DATE '2017-01-01'
  GROUP BY 1,2
  )
  WITH DATA
  UNIQUE PRIMARY INDEX (USER_ID, PROMO_DT)
  ON COMMIT PRESERVE ROWS
  ;
  "
  dbSendUpdate(dbCon,parRep(query))
  
  print(paste("round",k,as.Date(measureStart$measureStart) , "buy part"))
  
  # 2.38 DEMO  ----
  
  query = "
  CREATE MULTISET VOLATILE TABLE AU_DEMO AS (
  SELECT
  USR.USER_ID
  ,USR.PROMO_DT
  ,USR.PROMO_DT - DIM.USR_CREATED_DT AS TENURE_DAYS
  ,COALESCE(DIM.PAYPAL_LINK_STATE,-1) AS PAYPAL_LINK_STATE
  ,CASE WHEN DIM.PRDCTV_GNDR = 'M' THEN 0 
        WHEN DIM.PRDCTV_GNDR = 'F' THEN 1 ELSE -1 END AS PRDCTV_GNDR
  ,CASE WHEN DIM.APP_VISIT_IND = 'Y' THEN 1 
        WHEN DIM.APP_VISIT_IND = 'N' THEN 0 ELSE -1 END AS APP_VISIT_IND
  ,CASE WHEN EARLY.LIST_1ST_DT IS NULL OR EARLY.LIST_1ST_DT > USR.PROMO_DT THEN -1 ELSE USR.PROMO_DT - EARLY.LIST_1ST_DT END AS LIST_TENURE_DAYS
  ,CASE WHEN EARLY.SELL_1ST_DT IS NULL OR EARLY.SELL_1ST_DT > USR.PROMO_DT THEN -1 ELSE USR.PROMO_DT - EARLY.SELL_1ST_DT END AS SELL_TENURE_DAYS
  ,CASE WHEN EARLY.BUY_1ST_DT IS NULL OR EARLY.BUY_1ST_DT > USR.PROMO_DT THEN -1 ELSE USR.PROMO_DT - EARLY.BUY_1ST_DT END AS BUY_TENURE_DAYS
  FROM
  OPTIN90D_C2C USR
  LEFT JOIN
  PRS_SECURE_V.USER_DNA_DIM DIM
  ON USR.USER_ID = DIM.USER_ID
  LEFT JOIN
  PRS_RESTRICTED_V.USER_GXO_EARLY_ADPTN_CORE_MTRC EARLY
  ON DIM.PRIMARY_USER_ID = EARLY.PRIMARY_USER_ID
  )
  WITH DATA
  UNIQUE PRIMARY INDEX (USER_ID, PROMO_DT)
  ON COMMIT PRESERVE ROWS
  ;
  "
  dbSendUpdate(dbCon,parRep(query))
  
  print(paste("round",k,as.Date(measureStart$measureStart) , "demo part"))
  
  # 2.39 REVENUE  ----
  
  query = "
  CREATE MULTISET VOLATILE TABLE CK_TABLE AS 
  (
  SEL 
  CK.ITEM_ID
  , CK_TRANS_ID
  , CASE
  WHEN CK.SHPMT_MTHD_ID IN(15, 150, 151, 250, 311, 1650, 2350, 7115, 7750, 10151, 12350, 14650, 18650, 19350, 19606, 20101, 20309, 20508, 20701, 21101, 21250, 21601, 50312, 51550, 70509) THEN 1
  ELSE 0
  END AS LSTG_SHPG_LOCAL_ONLY_YN
  , CK_TRANS_DT         AS CK_DT         
  , LSTG_SITE_ID   
  , AUCT_START_DT 
  , SLR_ID
  , SOLD_QTY 
  , CASE
  WHEN SPS.SPS_SLR_LEVEL_CD = 1 THEN 'eTRS'
  WHEN SPS.SPS_SLR_LEVEL_CD = 2 THEN 'Above Standard'                            
  WHEN SPS.SPS_SLR_LEVEL_CD = 3 THEN 'Standard'     ---there is no Standard users since Sep. 2014
  WHEN SPS.SPS_SLR_LEVEL_CD = 4 THEN 'Below Standard'
  ELSE 'Unrated' END AS SLR_STANDARD
  , CASE
  WHEN STR.PROD_ID=3 THEN 'BASIC'
  WHEN STR.PROD_ID=4 THEN 'FEATURED'
  WHEN STR.PROD_ID=5 THEN 'ANCHOR'
  ELSE 'OTHERS' END AS SLR_SHOP
  , GMV_PLAN    AS GMV_USD_PLAN  
  
  FROM PRS_RESTRICTED_V.SLNG_TRANS_SUPER_FACT CK
  JOIN OPTIN90D_C2C U ON AUCT_END_DT >= DATE'2018-01-01' AND CK.SLR_ID = U.USER_ID
  LEFT JOIN PRS_RESTRICTED_V.SPS_LEVEL_METRIC_SUM  SPS ON SPS_PRGRM_ID=1 AND CK.SLR_ID = SPS.USER_ID
  AND CK_TRANS_DT BETWEEN SPS_SLR_LEVEL_SUM_START_DT  AND SPS_SLR_LEVEL_SUM_END_DT
  LEFT JOIN   DW_SUBSCRIPTION AS STR ON CK.SLR_ID=STR.ID
  AND (CAST(STR.END_DATE AS DATE)> CK_TRANS_DT OR STR.ACTIVE IN (0,1,2) OR CAST(END_DATE AS DATE) ='1969-12-31'  )      
  AND CAST(STR.SUB_DATE AS DATE) <= CK_TRANS_DT                  
  AND STR.PROD_ID IN (3,4,5) 
  )WITH DATA PRIMARY INDEX(ITEM_ID, CK_TRANS_ID)
  ON COMMIT PRESERVE ROWS
  ;
  "
  dbSendUpdate(dbCon,parRep(query))
  
  query = "
  CREATE MULTISET VOLATILE TABLE RVN_TMP AS
  (
  SELECT  CK_TRANS_ID
  ,LSTG_ID
  ,SLR_ID
  ,LKP.REV_BKT_ID
  , BLNG_CURNCY_ID
  ,SUM( CAST( -1 * AMT_BC AS FLOAT ) ) AMT_BC
  
  FROM     DW_GEM2_CMN_RVNU_I R
  JOIN DW_ACCT_ACTN_CODE_LKP LKP ON R.ACTN_CODE = LKP.ACTN_CODE   
  WHERE    ACCT_TRANS_DT  >= DATE'2018-01-01'  
  AND      LSTG_TYPE_CODE NOT IN( 10,15 )
  AND      ADJ_TYPE_ID NOT    IN( -1,-7,5 )
  AND      LSTG_SITE_ID <> 223
  AND SLR_ID IN (SEL USER_ID FROM OPTIN90D_C2C)
  GROUP BY 1,2,3,4,5
  
  UNION ALL
  
  /* Wash data of bad transactions */
  SELECT   CK_TRANS_ID
  ,LSTG_ID
  ,SLR_ID
  ,LKP.REV_BKT_ID
  , BLNG_CURNCY_ID
  ,SUM( CAST( AMT_BC AS FLOAT ) ) AMT_BC
  
  FROM     DW_GEM2_CMN_ADJ_RVNU_I R
  JOIN DW_ACCT_ACTN_CODE_LKP LKP ON R.ACTN_CODE = LKP.ACTN_CODE   
  WHERE    ACCT_TRANS_DT  >= DATE'2018-01-01' 
  AND      LSTG_TYPE_CODE NOT IN( 10,15 )
  AND      LSTG_SITE_ID <> 223
  AND      ISWACKO_YN_ID = 1
  AND SLR_ID IN (SEL USER_ID FROM OPTIN90D_C2C)
  GROUP BY 1,2,3,4,5
  )
  WITH DATA PRIMARY INDEX (
  CK_TRANS_ID
  ,LSTG_ID
  ,SLR_ID                 
  ,REV_BKT_ID
  ,BLNG_CURNCY_ID
  ) ON COMMIT PRESERVE ROWS
  ;
  "
  dbSendUpdate(dbCon,parRep(query))
  
  query = "
  CREATE MULTISET VOLATILE TABLE REV_TABLE AS
  (
  SEL
  CK_TRANS_ID
  , LSTG_ID
  , SUM(CASE WHEN REV_BKT_ID IN (29,30,31,32,33,34,35,36)THEN AMT_BC* (CAST(BPR.CURNCY_PLAN_RATE AS DECIMAL(18,2)))ELSE 0 END) AS REV_FEE_USD_PLAN
  , SUM(CASE WHEN REV_BKT_ID IN (29,30)  THEN AMT_BC* (CAST(BPR.CURNCY_PLAN_RATE AS DECIMAL(18,2))) ELSE 0 END) AS INS_FEE_USD_PLAN
  , SUM(CASE WHEN REV_BKT_ID IN (31,32)  THEN AMT_BC* (CAST(BPR.CURNCY_PLAN_RATE AS DECIMAL(18,2))) ELSE 0 END) AS FF_FEE_USD_PLAN
  , SUM(CASE WHEN REV_BKT_ID IN (33,34)  THEN AMT_BC* (CAST(BPR.CURNCY_PLAN_RATE AS DECIMAL(18,2))) ELSE 0 END) AS FVF_FEE_USD_PLAN
  
  FROM  RVN_TMP R
  LEFT JOIN  ACCESS_VIEWS.SSA_CURNCY_PLAN_RATE_DIM AS BPR  ON R.BLNG_CURNCY_ID = BPR.CURNCY_ID
  GROUP BY 1,2
  )WITH DATA PRIMARY INDEX(
  CK_TRANS_ID ,LSTG_ID)
  ON COMMIT PRESERVE ROWS;
  "
  dbSendUpdate(dbCon,parRep(query))
  
  query = "
  CREATE MULTISET VOLATILE TABLE FEE_DATA AS (
  SELECT
  CK.ITEM_ID
  ,CK.AUCT_START_DT
  ,CK.CK_DT
  ,CK.SLR_ID
  ,CK.SLR_STANDARD
  ,CK.SLR_SHOP
  ,CK.LSTG_SHPG_LOCAL_ONLY_YN
  ,SUM(ZEROIFNULL(CK.SOLD_QTY)) AS SI
  ,SUM(ZEROIFNULL(CK.GMV_USD_PLAN)) AS GMV_USD_PLAN -- CORRESPONDS TO GMV FIELD IN MOZART TABLE
  ,SUM(ZEROIFNULL(REV.FVF_FEE_USD_PLAN)) AS FVF_FEE_USD_PLAN -- CORRESPONDS TO FVF FIELD IN MOZART TABLE
  ,SUM(ZEROIFNULL(REV.INS_FEE_USD_PLAN)) AS INS_FEE_USD_PLAN 
  ,SUM(ZEROIFNULL(REV.FF_FEE_USD_PLAN)) AS FF_FEE_USD_PLAN 
  ,SUM(ZEROIFNULL(REV.REV_FEE_USD_PLAN)) AS REV_FEE_USD_PLAN 
  FROM
  CK_TABLE CK
  LEFT JOIN REV_TABLE REV ON CK.ITEM_ID = REV.LSTG_ID AND CK.CK_TRANS_ID = REV.CK_TRANS_ID
  GROUP BY 1,2,3,4,5,6,7
  )
  WITH DATA
  PRIMARY INDEX (ITEM_ID, CK_DT)
  ON COMMIT PRESERVE ROWS
  ;
  "
  dbSendUpdate(dbCon,parRep(query))
  
  
  query = "
  CREATE MULTISET VOLATILE TABLE AU_FEE_TXN AS (
  SELECT
  USR.USER_ID
  ,USR.PROMO_DT
  -- GMV PRE-PROMO
  ,SUM(CASE WHEN F.AUCT_START_DT BETWEEN USR.PROMO_DT - INTERVAL '30' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN GMV_USD_PLAN ELSE 0 END) AS T30_GMV_USD
  ,SUM(CASE WHEN F.AUCT_START_DT BETWEEN USR.PROMO_DT - INTERVAL '90' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN GMV_USD_PLAN ELSE 0 END) AS T90_GMV_USD
  ,SUM(CASE WHEN F.AUCT_START_DT BETWEEN USR.PROMO_DT - INTERVAL '180' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN GMV_USD_PLAN ELSE 0 END) AS T180_GMV_USD
  ,SUM(CASE WHEN F.AUCT_START_DT BETWEEN USR.PROMO_DT - INTERVAL '365' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN GMV_USD_PLAN ELSE 0 END) AS T365_GMV_USD
  -- FVF PRE-PROMO
  ,SUM(CASE WHEN F.AUCT_START_DT BETWEEN USR.PROMO_DT - INTERVAL '30' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN FVF_FEE_USD_PLAN ELSE 0 END) AS T30_FVF_USD
  ,SUM(CASE WHEN F.AUCT_START_DT BETWEEN USR.PROMO_DT - INTERVAL '90' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN FVF_FEE_USD_PLAN ELSE 0 END) AS T90_FVF_USD
  ,SUM(CASE WHEN F.AUCT_START_DT BETWEEN USR.PROMO_DT - INTERVAL '180' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN FVF_FEE_USD_PLAN ELSE 0 END) AS T180_FVF_USD
  ,SUM(CASE WHEN F.AUCT_START_DT BETWEEN USR.PROMO_DT - INTERVAL '365' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN FVF_FEE_USD_PLAN ELSE 0 END) AS T365_FVF_USD
  -- TAKE RATE PRE-PROMO
  ,CASE WHEN T30_GMV_USD > 0 THEN T30_FVF_USD / T30_GMV_USD ELSE 0 END AS T30_FVF_RATE
  ,CASE WHEN T90_GMV_USD > 0 THEN T90_FVF_USD / T90_GMV_USD ELSE 0 END AS T90_FVF_RATE
  ,CASE WHEN T180_GMV_USD > 0 THEN T180_FVF_USD / T180_GMV_USD ELSE 0 END AS T180_FVF_RATE
  ,CASE WHEN T365_GMV_USD > 0 THEN T365_FVF_USD / T365_GMV_USD ELSE 0 END AS T365_FVF_RATE
  
  -- POST GMV TO MATCH MEASUREMENT: INSERTED BETWEEN PROMO START AND END (~PROMO START + 4 DAYS), AND ADDITIONALLY CONVERTED IN NEXT 30 DAYS
  ,SUM(CASE WHEN F.AUCT_START_DT BETWEEN USR.PROMO_DT AND USR.PROMO_DT + INTERVAL '4' DAY AND 
                 F.CK_DT BETWEEN F.AUCT_START_DT AND F.AUCT_START_DT + INTERVAL '30' DAY THEN GMV_USD_PLAN ELSE 0 END) AS POST_PROMO_30D_GMV_USD
  -- PICKUP ONLY RATE
  ,SUM(CASE WHEN F.AUCT_START_DT BETWEEN USR.PROMO_DT - INTERVAL '30' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN LSTG_SHPG_LOCAL_ONLY_YN ELSE 0 END) AS T30_PICKUP_ONLY_CNT
  ,SUM(CASE WHEN F.AUCT_START_DT BETWEEN USR.PROMO_DT - INTERVAL '90' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN LSTG_SHPG_LOCAL_ONLY_YN ELSE 0 END) AS T90_PICKUP_ONLY_CNT
  ,SUM(CASE WHEN F.AUCT_START_DT BETWEEN USR.PROMO_DT - INTERVAL '180' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN LSTG_SHPG_LOCAL_ONLY_YN ELSE 0 END) AS T180_PICKUP_ONLY_CNT
  ,SUM(CASE WHEN F.AUCT_START_DT BETWEEN USR.PROMO_DT - INTERVAL '365' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN LSTG_SHPG_LOCAL_ONLY_YN ELSE 0 END) AS T365_PICKUP_ONLY_CNT
  ,SUM(CASE WHEN F.AUCT_START_DT BETWEEN USR.PROMO_DT - INTERVAL '30' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN 1 ELSE 0 END) AS T30_LSTG_CNT
  ,SUM(CASE WHEN F.AUCT_START_DT BETWEEN USR.PROMO_DT - INTERVAL '90' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN 1 ELSE 0 END) AS T90_LSTG_CNT
  ,SUM(CASE WHEN F.AUCT_START_DT BETWEEN USR.PROMO_DT - INTERVAL '180' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN 1 ELSE 0 END) AS T180_LSTG_CNT
  ,SUM(CASE WHEN F.AUCT_START_DT BETWEEN USR.PROMO_DT - INTERVAL '365' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN 1 ELSE 0 END) AS T365_LSTG_CNT
  ,CASE WHEN T30_LSTG_CNT > 0 THEN CAST(T30_PICKUP_ONLY_CNT AS DECIMAL(18,2)) / T30_LSTG_CNT ELSE 0 END AS T30_PICKUP_ONLY_RATE
  ,CASE WHEN T90_LSTG_CNT > 0 THEN CAST(T90_PICKUP_ONLY_CNT AS DECIMAL(18,2)) / T90_LSTG_CNT ELSE 0 END AS T90_PICKUP_ONLY_RATE
  ,CASE WHEN T180_LSTG_CNT > 0 THEN CAST(T180_PICKUP_ONLY_CNT AS DECIMAL(18,2)) / T180_LSTG_CNT ELSE 0 END AS T180_PICKUP_ONLY_RATE
  ,CASE WHEN T365_LSTG_CNT > 0 THEN CAST(T365_PICKUP_ONLY_CNT AS DECIMAL(18,2)) / T365_LSTG_CNT ELSE 0 END AS T365_PICKUP_ONLY_RATE
  -- MIN/MAX SELLER RATINGS OVER THE LAST YEAR
  ,MAX(CASE
  WHEN F.AUCT_START_DT BETWEEN USR.PROMO_DT - INTERVAL '365' DAY AND USR.PROMO_DT - INTERVAL '1' DAY AND SLR_STANDARD = 'ABOVE STANDARD' THEN 1
  WHEN F.AUCT_START_DT BETWEEN USR.PROMO_DT - INTERVAL '365' DAY AND USR.PROMO_DT - INTERVAL '1' DAY AND SLR_STANDARD = 'BELOW STANDARD' THEN -1
  WHEN F.AUCT_START_DT BETWEEN USR.PROMO_DT - INTERVAL '365' DAY AND USR.PROMO_DT - INTERVAL '1' DAY AND SLR_STANDARD = 'ETRS' THEN 2
  ELSE 0
  END) AS MAX_T365_SLR_STANDARD
  ,MIN(CASE
  WHEN F.AUCT_START_DT BETWEEN USR.PROMO_DT - INTERVAL '365' DAY AND USR.PROMO_DT - INTERVAL '1' DAY AND SLR_STANDARD = 'ABOVE STANDARD' THEN 1
  WHEN F.AUCT_START_DT BETWEEN USR.PROMO_DT - INTERVAL '365' DAY AND USR.PROMO_DT - INTERVAL '1' DAY AND SLR_STANDARD = 'BELOW STANDARD' THEN -1
  WHEN F.AUCT_START_DT BETWEEN USR.PROMO_DT - INTERVAL '365' DAY AND USR.PROMO_DT - INTERVAL '1' DAY AND SLR_STANDARD = 'ETRS' THEN 2
  ELSE 0
  END) AS MIN_T365_SLR_STANDARD
  -- RANKING ON SHOP (ANCHOR, FEATURED, BASIC, NO SHOP)
  ,MAX(CASE
  WHEN F.AUCT_START_DT BETWEEN USR.PROMO_DT - INTERVAL '365' DAY AND USR.PROMO_DT - INTERVAL '1' DAY AND SLR_SHOP = 'ANCHOR' THEN 3
  WHEN F.AUCT_START_DT BETWEEN USR.PROMO_DT - INTERVAL '365' DAY AND USR.PROMO_DT - INTERVAL '1' DAY AND SLR_SHOP = 'FEATURED' THEN 2
  WHEN F.AUCT_START_DT BETWEEN USR.PROMO_DT - INTERVAL '365' DAY AND USR.PROMO_DT - INTERVAL '1' DAY AND SLR_SHOP = 'BASIC' THEN 1
  ELSE 0
  END) AS MAX_T365_SHOP_LEVEL
  FROM
  OPTIN90D_C2C USR
  LEFT JOIN
  FEE_DATA F
  ON USR.USER_ID = F.SLR_ID
  AND F.AUCT_START_DT BETWEEN USR.PROMO_DT - INTERVAL '365' DAY AND USR.PROMO_DT + INTERVAL '4' DAY
  GROUP BY 1,2
  )
  WITH DATA
  UNIQUE PRIMARY INDEX (USER_ID, PROMO_DT)
  ON COMMIT PRESERVE ROWS
  ;
  "
  dbSendUpdate(dbCon,parRep(query))
  
  print(paste("round",k,as.Date(measureStart$measureStart) , "revenue part"))
  
  # 2.310 BBE  ----
  
  query = "
  CREATE MULTISET VOLATILE TABLE BBE_DATA AS (
  SELECT
  DEF.SLR_ID
  ,DEF.TRANS_DT
  ,SUM(CASE
  WHEN DEF.ESC_SNAD_FLAG = 1 THEN CAST(1 AS DECIMAL(18,2))
  WHEN DEF.OPEN_SNAD_FLAG = 1 THEN CAST(1 AS DECIMAL(18,2))
  WHEN DEF.RTRN_SNAD_FLAG = 1 THEN CAST(1 AS DECIMAL(18,2))
  WHEN DEF.STOCKOUT_FLAG = 1 THEN CAST(1 AS DECIMAL(18,2))
  WHEN DEF.ESC_INR_FLAG = 1 THEN CAST(1 AS DECIMAL(18,2))
  WHEN DEF.OPEN_INR_FLAG = 1 THEN CAST(1 AS DECIMAL(18,2))
  WHEN DEF.LOW_DSR_IAD_FLAG = 1 THEN CAST(1 AS DECIMAL(18,2))
  WHEN DEF.BYR_TO_SLR_NN_FLAG = 1 THEN CAST(1 AS DECIMAL(18,2))
  WHEN DEF.SNAD_MSG_FLAG = 1 THEN
  CASE
  WHEN DEF.TRANS_SITE_ID = 0 THEN CAST(0.90 AS DECIMAL(18,2))
  WHEN DEF.TRANS_SITE_ID = 3 THEN CAST(0.90 AS DECIMAL(18,2))
  WHEN DEF.TRANS_SITE_ID = 77 THEN CAST(0.65 AS DECIMAL(18,2))
  ELSE CAST(1 AS DECIMAL(18,2))
  END
  WHEN DEF.INR_MSG_FLAG = 1 THEN
  CASE
  WHEN DEF.TRANS_SITE_ID = 0 THEN CAST(0.90 AS DECIMAL(18,2))
  WHEN DEF.TRANS_SITE_ID = 3 THEN CAST(0.90 AS DECIMAL(18,2))
  WHEN DEF.TRANS_SITE_ID = 77 THEN CAST(0.65 AS DECIMAL(18,2))
  ELSE CAST(1 AS DECIMAL(18,2))
  END
  ELSE CAST(0 AS DECIMAL(18,2))
  END) AS DEFECT_CNT
  ,COUNT(1) AS TRANS_CNT
  
  FROM
  PRS_RESTRICTED_V.EBAY_TRANS_RLTD_EVENT DEF
  WHERE
  -- OUR DATA SET CONTAINS PROMOS FROM JAN 2019 ONWARD, AND WE WANT TO LOOK BACK IN HISTORY FOR ABOUT A YEAR.
  DEF.TRANS_DT >= '2018-01-01'
  AND DEF.BBE_ELGB_TRANS_IND = 1
  AND DEF.SLR_ID IN (SELECT DISTINCT USER_ID FROM OPTIN90D_C2C)
  GROUP BY 1,2
  )
  WITH DATA
  UNIQUE PRIMARY INDEX (SLR_ID, TRANS_DT)
  ON COMMIT PRESERVE ROWS
  ;
  "
  dbSendUpdate(dbCon,parRep(query))
  
  query = "
  CREATE MULTISET VOLATILE TABLE AU_BBE AS (
  SELECT
  USR.USER_ID
  ,USR.PROMO_DT
  ,SUM(CASE WHEN B.TRANS_DT BETWEEN USR.PROMO_DT - INTERVAL '30' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN DEFECT_CNT ELSE 0 END) AS T30_DEFECT_CNT
  ,SUM(CASE WHEN B.TRANS_DT BETWEEN USR.PROMO_DT - INTERVAL '90' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN DEFECT_CNT ELSE 0 END) AS T90_DEFECT_CNT
  ,SUM(CASE WHEN B.TRANS_DT BETWEEN USR.PROMO_DT - INTERVAL '180' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN DEFECT_CNT ELSE 0 END) AS T180_DEFECT_CNT
  ,SUM(CASE WHEN B.TRANS_DT BETWEEN USR.PROMO_DT - INTERVAL '365' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN DEFECT_CNT ELSE 0 END) AS T365_DEFECT_CNT
  ,SUM(CASE WHEN B.TRANS_DT BETWEEN USR.PROMO_DT - INTERVAL '30' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN TRANS_CNT ELSE 0 END) AS T30_TRANS_CNT
  ,SUM(CASE WHEN B.TRANS_DT BETWEEN USR.PROMO_DT - INTERVAL '90' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN TRANS_CNT ELSE 0 END) AS T90_TRANS_CNT
  ,SUM(CASE WHEN B.TRANS_DT BETWEEN USR.PROMO_DT - INTERVAL '180' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN TRANS_CNT ELSE 0 END) AS T180_TRANS_CNT
  ,SUM(CASE WHEN B.TRANS_DT BETWEEN USR.PROMO_DT - INTERVAL '365' DAY AND USR.PROMO_DT - INTERVAL '1' DAY THEN TRANS_CNT ELSE 0 END) AS T365_TRANS_CNT
  ,CASE WHEN T30_TRANS_CNT > 0 THEN  CAST(T30_DEFECT_CNT AS DECIMAL(18,4)) / T30_TRANS_CNT ELSE 0 END AS T30_DEFECT_RATE
  ,CASE WHEN T90_TRANS_CNT > 0 THEN  CAST(T90_DEFECT_CNT AS DECIMAL(18,4)) / T90_TRANS_CNT ELSE 0 END AS T90_DEFECT_RATE
  ,CASE WHEN T180_TRANS_CNT > 0 THEN  CAST(T180_DEFECT_CNT AS DECIMAL(18,4)) / T180_TRANS_CNT ELSE 0 END AS T180_DEFECT_RATE
  ,CASE WHEN T365_TRANS_CNT > 0 THEN  CAST(T365_DEFECT_CNT AS DECIMAL(18,4)) / T365_TRANS_CNT ELSE 0 END AS T365_DEFECT_RATE
  FROM
  OPTIN90D_C2C USR
  LEFT JOIN
  BBE_DATA B
  ON USR.USER_ID = B.SLR_ID
  AND B.TRANS_DT BETWEEN USR.PROMO_DT - INTERVAL '365' DAY AND USR.PROMO_DT - INTERVAL '1' DAY
  GROUP BY 1,2
  )
  WITH DATA
  UNIQUE PRIMARY INDEX (USER_ID, PROMO_DT)
  ON COMMIT PRESERVE ROWS
  ;
  "
  dbSendUpdate(dbCon,parRep(query))
  
  print(paste("round",k,as.Date(measureStart$measureStart) , "bbe part"))
  
  # 2.4 COMBINE  ----
  # dbSendUpdate(dbCon, "drop table au_c2c_fulldataset")
  
  query = "
  create multiset volatile table au_c2c_fulldataset as (
  select
  usr.user_id
  ,usr.promo_dt
  -- add: binary t/c indicator
  ,case
  when usr.tc_segment = 'T' then 1
  when usr.tc_segment = 'C' then 0
  else -1 end as tc_ind
  , 'other' as segment
  ,usr.promo_optin_ind
  -- ===========================
  -- add: binary response indicator
  ,resp.promo_lstg_cnt_2d
  ,resp.promo_lstg_cnt_7d
  ,resp.promo_gmv_2d
  ,resp.promo_gmv_7d
  ,resp.promo_si_2d
  ,resp.promo_si_7d
  -- ===========================
  ,optin.prev_optin_cnt
  -- ===========================
  ,lst.t30_lstg_cnt
  ,lst.t90_lstg_cnt
  ,lst.t180_lstg_cnt
  ,lst.t365_lstg_cnt
  ,lst.tprev30_lstg_cnt
  ,lst.tprev90_lstg_cnt	
  ,lst.tprev180_lstg_cnt
  ,lst.tprev365_lstg_cnt
  ,lst.t30_lstg_rate
  ,lst.t90_lstg_rate
  ,lst.t180_lstg_rate
  ,lst.t365_lstg_rate
  ,lst.lstg_tool_pro_ind
  ,lst.lstg_tool_ios_ind
  ,lst.lstg_tool_android_ind
  ,lst.lstg_tool_mob_ind
  -- ===========================
  ,vert.t365_vrtcl_cnt
  -- ===========================
  ,sell.t30_si_cnt
  ,sell.t90_si_cnt
  ,sell.t180_si_cnt
  ,sell.t365_si_cnt
  ,sell.tprev30_si_cnt
  ,sell.tprev90_si_cnt
  ,sell.tprev180_si_cnt
  ,sell.tprev365_si_cnt
  ,sell.t30_si_rate
  ,sell.t90_si_rate
  ,sell.t180_si_rate
  ,sell.t365_si_rate
  ,sell.t30_gmv_amt
  ,sell.t90_gmv_amt
  ,sell.t180_gmv_amt
  ,sell.t365_gmv_amt
  ,sell.tprev30_gmv_amt
  ,sell.tprev90_gmv_amt
  ,sell.tprev180_gmv_amt
  ,sell.tprev365_gmv_amt
  ,sell.t30_gmv_rate
  ,sell.t90_gmv_rate
  ,sell.t180_gmv_rate
  ,sell.t365_gmv_rate
  ,sell.t30_asp
  ,sell.t90_asp
  ,sell.t180_asp
  ,sell.t365_asp
  -- add: success rate of sells/listings (note: for multiple sells for a single listing this may become greater than 1)
  ,case when lst.t30_lstg_cnt = 0 then 0 else cast(sell.t30_si_cnt as decimal(18,2)) / lst.t30_lstg_cnt end as t30_lstg_success_rate
  ,case when lst.t90_lstg_cnt = 0 then 0 else cast(sell.t90_si_cnt as decimal(18,2)) / lst.t90_lstg_cnt end as t90_lstg_success_rate
  ,case when lst.t180_lstg_cnt = 0 then 0 else cast(sell.t180_si_cnt as decimal(18,2)) / lst.t180_lstg_cnt end as t180_lstg_success_rate
  -- ===========================
  ,email.t90_email_sent_cnt
  ,email.tprev90_email_sent_cnt
  ,email.t30_email_sent_cnt
  ,email.tprev30_email_sent_cnt
  ,email.t90_email_open_cnt
  ,email.tprev90_email_open_cnt
  ,email.t30_email_open_cnt
  ,email.tprev30_email_open_cnt
  ,email.t90_email_click_cnt
  ,email.tprev90_email_click_cnt
  ,email.t30_email_click_cnt
  ,email.tprev30_email_click_cnt
  ,email.t90_email_or
  ,email.t30_email_or
  ,email.t90_email_ctor
  ,email.t30_email_ctor
  ,email.t90_email_opens_rate
  ,email.t30_email_opens_rate
  ,email.t90_email_clicks_rate
  ,email.t30_email_clicks_rate
  -- ===========================
  ,sb.t90_vi_cnt
  ,sb.tprev90_vi_cnt
  ,sb.t30_vi_cnt
  ,sb.tprev30_vi_cnt
  ,sb.t90_srch_cnt
  ,sb.tprev90_srch_cnt
  ,sb.t30_srch_cnt
  ,sb.tprev30_srch_cnt
  ,sb.t90_vi_rate
  ,sb.t30_vi_rate
  ,sb.t90_srch_rate
  ,sb.t30_srch_rate
  -- ===========================
  ,buy.t365_prchs_cnt
  ,buy.tprev365_prchs_cnt
  ,buy.t90_prchs_cnt
  ,buy.tprev90_prchs_cnt
  ,buy.t30_prchs_cnt
  ,buy.tprev30_prchs_cnt
  ,buy.t365_gmb_amt
  ,buy.tprev365_gmb_amt
  ,buy.t90_gmb_amt
  ,buy.tprev90_gmb_amt
  ,buy.t30_gmb_amt
  ,buy.tprev30_gmb_amt
  ,buy.t365_prchs_rate
  ,buy.t90_prchs_rate
  ,buy.t30_prchs_rate
  ,buy.t365_gmb_rate
  ,buy.t90_gmb_rate
  ,buy.t30_gmb_rate
  -- add: sell/buy ratios
  ,case when buy.t365_prchs_cnt = 0 then -1 else cast(sell.t365_si_cnt as decimal(18,2)) / buy.t365_prchs_cnt end as t365_sbi_ratio
  ,case when buy.t365_gmb_amt = 0 then -1 else cast(sell.t365_gmv_amt as decimal(18,2)) / buy.t365_gmb_amt end as t365_gmvb_ratio
  -- ===========================
  ,demo.tenure_days
  ,demo.paypal_link_state
  ,demo.prdctv_gndr
  ,demo.app_visit_ind
  ,demo.list_tenure_days
  ,demo.sell_tenure_days
  ,demo.buy_tenure_days
  -- ===========================
  ,fee.t30_fvf_rate
  ,fee.t90_fvf_rate
  ,fee.t180_fvf_rate
  ,fee.t365_fvf_rate
  ,fee.post_promo_30d_gmv_usd
  ,fee.t30_pickup_only_rate
  ,fee.t90_pickup_only_rate
  ,fee.t180_pickup_only_rate
  ,fee.t365_pickup_only_rate
  ,fee.max_t365_slr_standard
  ,fee.min_t365_slr_standard
  ,fee.max_t365_shop_level
  -- ===========================
  -- add bbe indicator
  ,case when bbe.t30_defect_cnt > 0 then 1 else 0 end as t30_defect_ind
  ,case when bbe.t90_defect_cnt > 0 then 1 else 0 end as t90_defect_ind
  ,case when bbe.t180_defect_cnt > 0 then 1 else 0 end as t180_defect_ind
  ,case when bbe.t365_defect_cnt > 0 then 1 else 0 end as t365_defect_ind
  ,bbe.t30_defect_rate
  ,bbe.t90_defect_rate
  ,bbe.t180_defect_rate
  ,bbe.t365_defect_rate
  from 		optin90d_c2c usr
  inner join	au_response		resp	on usr.user_id = resp.user_id	and usr.promo_dt = resp.promo_dt
  inner join	au_prev_optins optin	on usr.user_id = optin.user_id	and usr.promo_dt = optin.promo_dt
  inner join	au_listings 	lst		on usr.user_id = lst.user_id	and usr.promo_dt = lst.promo_dt
  inner join	au_verticals	vert	on usr.user_id = vert.user_id	and usr.promo_dt = vert.promo_dt
  inner join	au_sells	    sell	on usr.user_id = sell.user_id	and usr.promo_dt = sell.promo_dt
  inner join	au_email	    email	on usr.user_id = email.user_id	and usr.promo_dt = email.promo_dt
  inner join	au_search_browse	sb		on usr.user_id = sb.user_id		and usr.promo_dt = sb.promo_dt
  inner join	au_buys 	    buy		on usr.user_id = buy.user_id	and usr.promo_dt = buy.promo_dt
  inner join	au_demo		    demo	on usr.user_id = demo.user_id	and usr.promo_dt = demo.promo_dt
  inner join  au_fee_txn		fee 	on usr.user_id = fee.user_id	and usr.promo_dt = fee.promo_dt
  inner join  au_bbe				bbe		on usr.user_id = bbe.user_id	and usr.promo_dt = bbe.promo_dt
  )
  with data
  unique primary index (user_id, promo_dt)
  on commit preserve rows
  ;
  "
  dbSendUpdate(dbCon,parRep(query))
  
  print(paste("round",k,as.Date(measureStart$measureStart) , "combine part"))
  
  # 2.5 ENGAGED ONLY and save data ----
  
  query = "
  CREATE MULTISET VOLATILE TABLE AU_LSTG_MNTHS AS (
  SELECT
  USR.USER_ID
  ,USR.PROMO_DT
  ,COUNT(DISTINCT CASE WHEN SD.CAL_DT >= USR.PROMO_DT - INTERVAL '365' DAY AND NEW_LSTG_CNT>0 THEN EXTRACT(MONTH FROM CAL_DT)  END) AS LSTG_MONTHS
  FROM
  OPTIN90D_C2C USR
  LEFT JOIN
  PRS_RESTRICTED_V.SLR_INVNTRY_SD SD
  ON USR.USER_ID = SD.SLR_ID
  AND SD.CAL_DT BETWEEN USR.PROMO_DT - INTERVAL '730' DAY AND  USR.PROMO_DT - INTERVAL '1' DAY
  AND SD.CAL_DT >= DATE '2017-01-01'
  AND SD.LSTG_SITE_ID = 15	
  GROUP BY 1,2
  )
  WITH DATA
  UNIQUE PRIMARY INDEX (USER_ID, PROMO_DT)
  ON COMMIT PRESERVE ROWS
  ;
  "
dbSendUpdate(dbCon,parRep(query))

# dbSendUpdate(dbCon, "drop table AU_C2C_0_FULLDATASET")

query = "
CREATE MULTISET VOLATILE TABLE AU_C2C_0_FULLDATASET AS (
SELECT 
USR.*
,LM.LSTG_MONTHS  AS lstg_months
FROM 	AU_C2C_FULLDATASET USR
INNER JOIN	AU_LSTG_MNTHS LM	
ON USR.USER_ID = LM.USER_ID	AND USR.PROMO_DT = LM.PROMO_DT
)
WITH DATA
UNIQUE PRIMARY INDEX (USER_ID, PROMO_DT)
ON COMMIT PRESERVE ROWS
;
"
dbSendUpdate(dbCon,parRep(query))

print(paste("round",k,as.Date(measureStart$measureStart) , "engaged only part"))

Engagedata = dbGetQuery(dbCon, "SELECT * FROM AU_C2C_0_FULLDATASET WHERE T365_LSTG_CNT>5 AND LSTG_MONTHS>=4;")
write.csv(Engagedata,paste("AU_C2C_ENGAGED_",as.Date(measureStart$measureStart),".csv",sep=""),row.names = FALSE)

print(paste("################### round",k,as.Date(measureStart$measureStart) , "end ###################"))

invisible(dbDisconnect(dbCon)) 
}
